move msal config into session folder
